move krt utils to usecase pacakge
